+++ maxbunny/tests/mock_http.py
@@ -34,6 +34,11 @@ def http_mock_grant_subscription_permission(uri='tests.local', status_code=201):
         status=status_code,
     )
 
+def http_mock_delete_token(uri='tests.local', status_code=204):
+    httpretty.register_uri(
+        httpretty.DELETE, re.compile("http://{}/tokens/\w+".format(uri)),
+        status=status_code,
+    )
 
 def http_mock_subscribe_user(uri='tests.local', fail_response=None):
     responses = [httpretty.Response(
@@ -44,7 +49,7 @@ def http_mock_subscribe_user(uri='tests.local', fail_response=None):
         responses.insert(0, fail_response)
 
     httpretty.register_uri(
-        httpretty.POST, re.compile("http://{}/people/\w+/subscriptions".format(uri)),
+        httpretty.POST, re.compile("http://{}/contexts/\w+/subscriptions".format(uri)),
         responses=responses
     )